socketbox-preload: Deny listen() for sockets with a port of 0
The issue here surrounds daemons that do not call bind() directly,
but spawn an external agent to do so instead (think if authbind
support for a particular program was integrated directly into the
program itself instead of with the included LD_PRELOAD library)
If an external agent were to perform the bind(), then the replace
operation would only occur within the context of the agent, rather
than in the original program. This would result in the supposed
bind() operation acting as a no-op. If the agent program were to
exit with a status of 0, and the original program took that as a
success, then it would be indistinguishable from if the bind were
to be actually successful. Then, when the program calls listen(),
it would unknowingly result in binding to a random port number
on 0.0.0.0 (IPv4) or :: (IPv6), and this could result in
unauthorized access to the service (in the absence of a firewall).

This modification fixes that, by blocking listen() if the socket
is not yet bound to an address. This fixes the above problem
nicely, at the expense of breaking any program that does not call
bind() before listen() (not likely unless there was an actual
intent to listen on a random port on :: or 0.0.0.0 [e.g. some
FTP servers or P2P clients])